requirements.txt
fastapi==0.112.1
uvicorn==0.30.6
requests==2.32.3
yt-dlp==2024.8.6
torch==2.0.1
torchaudio
torchvision
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
numpy==1.23.5
pandas==2.2.2
python-dotenv==1.0.1
python-multipart==0.0.9
beautifulsoup4==4.12.3
selenium==4.23.1
undetected-chromedriver==3.5.5
kiteconnect==5.0.1
PyMuPDF==1.24.9
openai==1.41.1
google-generativeai==0.7.2
alembic==1.13.2
SQLAlchemy==2.0.35
psycopg2==2.9.9
pydantic-settings==2.5.2
passlib==1.7.4
python-jose==3.3.0
fastapi-mail==1.4.1
inflect==7.4.0
jwt==1.3.1
langchain==0.3.3
pinecone-client==5.0.1
firebase-admin==6.5.0

